revision: invert class dependency
