+++ lib/Plugins/Features/ReviewImproved/Decorator/CatDecorator.php
@@ -2,6 +2,18 @@
 
 namespace Features\ReviewImproved\Decorator ;
 
-use Features\AbstractCatDecorator;
 
-class CatDecorator extends AbstractCatDecorator { }
+class CatDecorator extends \Features\ReviewExtended\Decorator\CatDecorator {
+
+    public function decorate() {
+        parent::decorate();
+
+        $this->template->review_type = 'improved';
+
+        if ( $this->controller->isRevision() ) {
+            $this->template->showReplaceOptionsInSearch = false ;
+        }
+
+    }
+
+}
